@@ -20,6 +20,6 @@
         var _VAULT_GLOBALS = @Html.Raw(Model.AsJson());
     </script>
 
-    <script src="~/js/dist/Index.min.js" asp-append-version="true" integrity="sha512-hx49T/WgDWXL9PWaIlkM/pATb/j60NvyEZStZhH1qS8Wkl8d2+qPbxhVEb0U8GrptV+KGJf2JCLC++Z1j8e76A=="></script>
+    <script src="~/js/dist/Index.min.js" asp-append-version="true" integrity="sha512-8ni6tCdvJN9w0ilCAW3GJsa1yt/6M6oU7aPzRuXdS4GKtIl7/cI6PDELg+Lcqx7nDXCkTadP3X8/tbg2Zf1jrw=="></script>
 
 }
